refactor(TableGroupValidator.java) : validate 파일 validator 패키지 이동
