+++ module-product/src/main/java/kitchenpos/product/validator/MenuProductValidatorImpl.java
@@ -1,6 +1,7 @@
 package kitchenpos.product.validator;
 
 import kitchenpos.menu.domain.*;
+import kitchenpos.menu.validator.MenuProductValidator;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.repository.ProductRepository;
 import org.springframework.stereotype.Component;
