+++ module-table/src/main/java/kitchenpos/table/application/TableGroupService.java
@@ -7,6 +7,7 @@ import kitchenpos.table.dto.CreateTableGroupRequest;
 import kitchenpos.table.dto.TableGroupResponse;
 import kitchenpos.table.repository.OrderTableRepository;
 import kitchenpos.table.repository.TableGroupRepository;
+import kitchenpos.table.validator.TableGroupValidator;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
