+++ module-table/src/main/java/kitchenpos/table/validator/TableGroupValidator.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.application;
+package kitchenpos.table.validator;
 
 import kitchenpos.table.domain.TableGroup;
